Remove old subgraph folder
