--- docker-compose.env
@@ -1,4 +1,0 @@
-# used by hardhat
-ETH1_HTTP_PORT=9545
-# used by hardhat and integration tests
-MNEMONIC='tattoo bulk finish hurdle bread monster tip say absent cake high damp'
