--- new-subgraph/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "the-creaton",
-  "license": "UNLICENSED",
-  "scripts": {
-    "codegen": "graph codegen",
-    "build": "graph build",
-    "deploy": "graph deploy --node https://api.thegraph.com/deploy/ --ipfs https://api.thegraph.com/ipfs/ kdd-anon-user/the-creaton",
-    "create-local": "graph create --node http://localhost:8020/ kdd-anon-user/the-creaton",
-    "remove-local": "graph remove --node http://localhost:8020/ kdd-anon-user/the-creaton",
-    "deploy-local": "graph deploy --node http://localhost:8020/ --ipfs http://localhost:5001 kdd-anon-user/the-creaton"
-  },
-  "dependencies": {
-    "@graphprotocol/graph-cli": "0.19.0",
-    "@graphprotocol/graph-ts": "0.19.0"
-  }
-}
