--- new-subgraph/schema.graphql
@@ -1,30 +1,0 @@
-type Creator @entity {
-  id: ID!
-  user: Bytes!
-  creatorContract: Bytes!
-  description: String!
-  subscriptionPrice: BigInt!
-  timestamp: BigInt!
-  subscribers: [Subscriber!]! @derivedFrom(field: "creator")
-}
-
-type Subscriber @entity {
-  id: ID!
-  creator: Creator!
-  creatorContract: Bytes!
-  user: Bytes!
-  sig_key: String!
-  pub_key: String!
-  status: String!
-}
-
-type Content @entity {
-  id: ID!
-  creator: Creator!
-  creatorContract: Bytes!
-  name: String!
-  type: String!
-  description: String!
-  date: String!
-  ipfs: String!
-}
