+++ subgraph/docker-compose.yml
@@ -1,7 +1,7 @@
 version: '3'
 services:
   graph-node:
-    image: graphprotocol/graph-node:v0.20.0
+    image: graphprotocol/graph-node
     ports:
       - '8000:8000'
       - '8001:8001'
@@ -18,16 +18,16 @@ services:
       postgres_db: graph-node
       ipfs: 'ipfs:5001'
       ethereum: 'mainnet:http://host.docker.internal:8545'
-      # RUST_LOG: info
-      # RUST_BACKTRACE: full
+      RUST_LOG: info
+      RUST_BACKTRACE: full
   ipfs:
-    image: ipfs/go-ipfs:release
+    image: ipfs/go-ipfs
     ports:
       - '5001:5001'
     environment:
       IPFS_PATH: /tmp
   postgres:
-    image: postgres:9.6.19
+    image: postgres
     ports:
       - '5432:5432'
     command: ['postgres', '-cshared_preload_libraries=pg_stat_statements']
