+++ subgraph/package.json
@@ -1,35 +1,16 @@
 {
-  "name": "creaton-subgraph",
-  "description": "Subgraph (thegraph.com) for Creaton",
-  "private": true,
-  "version": "0.0.1",
-  "devDependencies": {
-    "@graphprotocol/graph-cli": "^0.19.0",
-    "@graphprotocol/graph-ts": "^0.19.0",
-    "chokidar": "^3.4.2",
-    "cross-var": "^1.1.0",
-    "dotenv-cli": "^3.1.0",
-    "fs-extra": "^9.0.0",
-    "handlebars": "^4.7.6",
-    "npm-run-all": "^4.1.5",
-    "wait-on": "^4.0.1"
-  },
+  "name": "the-creaton",
+  "license": "UNLICENSED",
   "scripts": {
-    "local:create:silent": "cross-var graph create $SUBGRAPH_NAME --node http://127.0.0.1:8020 || true",
-    "local:setup": "wait-on http-get://127.0.0.1:8000 && npm run local:create:silent",
-    "codegen": "rm -Rf generated && graph codegen",
+    "codegen": "graph codegen",
     "build": "graph build",
-    "generate": "node scripts/generate.js",
-    "watch": "node scripts/watch.js",
-    "dev": "npm-run-all \"local:deploy -- {1}\" \"watch -- {1}\" --",
-    "local:redeploy": "graph codegen && wait-on http-get://127.0.0.1:8000 && npm run local:redeploy:1",
-    "local:redeploy:1": "dotenv -- cross-var graph deploy $SUBGRAPH_NAME --ipfs http://localhost:5001 --node http://127.0.0.1:8020 --debug",
-    "local:deploy": "npm-run-all \"generate -- {1} mainnet\" local:redeploy --",
-    "mainnet:redeploy": "dotenv -- cross-var graph deploy --access-token $THEGRAPH_TOKEN $SUBGRAPH_NAME --ipfs https://api.thegraph.com/ipfs/ --node https://api.thegraph.com/deploy/ --debug",
-    "mainnet:deploy": "npm-run-all \"generate -- {1} mainnet\" mainnet:redeploy --",
-    "staging:redeploy": "dotenv -- cross-var graph deploy --access-token $THEGRAPH_TOKEN $SUBGRAPH_NAME --ipfs https://api.thegraph.com/ipfs/ --node https://api.thegraph.com/deploy/ --debug",
-    "staging:deploy": "npm-run-all \"generate -- {1} goerli\" staging:redeploy --",
-    "kovan:redeploy": "dotenv -- cross-var graph deploy --access-token $THEGRAPH_TOKEN $SUBGRAPH_NAME --ipfs https://api.thegraph.com/ipfs/ --node https://api.thegraph.com/deploy/ --debug",
-    "kovan:deploy": "npm-run-all \"generate -- {1} kovan\" kovan:redeploy --"
+    "deploy": "graph deploy --node https://api.thegraph.com/deploy/ --ipfs https://api.thegraph.com/ipfs/ creation-io/creaton",
+    "create-local": "graph create --node http://localhost:8020/ creation-io/creaton",
+    "remove-local": "graph remove --node http://localhost:8020/ creation-io/creaton",
+    "deploy-local": "graph deploy --node http://localhost:8020/ --ipfs http://localhost:5001 creation-io/creaton"
+  },
+  "dependencies": {
+    "@graphprotocol/graph-cli": "0.19.0",
+    "@graphprotocol/graph-ts": "0.19.0"
   }
 }
