+++ subgraph/schema.graphql
@@ -2,8 +2,29 @@ type Creator @entity {
   id: ID!
   user: Bytes!
   creatorContract: Bytes!
-  title: String!
+  description: String!
   subscriptionPrice: BigInt!
-  avatarURL: String!
   timestamp: BigInt!
+  subscribers: [Subscriber!]! @derivedFrom(field: "creator")
+}
+
+type Subscriber @entity {
+  id: ID!
+  creator: Creator!
+  creatorContract: Bytes!
+  user: Bytes!
+  sig_key: String!
+  pub_key: String!
+  status: String!
+}
+
+type Content @entity {
+  id: ID!
+  creator: Creator!
+  creatorContract: Bytes!
+  name: String!
+  type: String!
+  description: String!
+  date: String!
+  ipfs: String!
 }
