--- subgraph/scripts/.eslintrc.js
@@ -1,6 +1,0 @@
-module.exports = {
-  parser: 'esprima',
-  rules: {
-    semi: 'error',
-  },
-};
