--- subgraph/templates/subgraph.yaml
@@ -1,25 +1,0 @@
-specVersion: 0.0.2
-description: Creaton
-schema:
-  file: ./schema.graphql
-dataSources:
-  - kind: ethereum/contract
-    name: CreatonFactory
-    network: {{chainName}}
-    source:
-      address: '{{contracts.CreatonFactory.address}}'
-      abi: CreatonFactory
-      startBlock: {{contracts.CreatonFactory.receipt.blockNumber}}
-    mapping:
-      kind: ethereum/events
-      apiVersion: 0.0.3
-      language: wasm/assemblyscript
-      entities:
-        - Creator
-      abis:
-        - name: CreatonFactory
-          file: ./abis/CreatonFactory.json
-      eventHandlers:
-        - event: CreatorDeployed(address,address,string,string,uint256)
-          handler: handleCreatorDeployed
-      file: ./src/mapping.ts
